FDG-6275 moved relevant accordion tests
